backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pymongo==4.6.0
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6

# Kafka integration
kafka-python==2.0.2
aiokafka==0.9.0

# HTTP client and utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Image and media processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Geospatial and location
geopy==2.4.1
shapely==2.0.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# WebSocket support
websockets==12.0

# Email and SMS
aiosmtplib==3.0.1
twilio==8.10.3

# File upload and storage
boto3==1.34.0  # For AWS S3 integration

# Rate limiting
slowapi==0.1.9

# CORS and middleware
fastapi-cors==0.0.6